feat(utils): add number utils
